+++ CHANGELOG.md
@@ -36,6 +36,7 @@ speaking, considered to be the player's experience.
 - Select sortied units only if the Fort is still selected, instead of overriding
   the selection, preventing accidentally giving orders to the wrong units in
   multiplayer.
+- Made development slightly easier.
 
 
 ## [3.2.11] — 2026-08-24